changes: moved dto to infrastructure layer
