+++ backend/src/infrastructure/dto/business/business-response.dto.ts
