+++ backend/src/infrastructure/dto/business/search-business.dto.ts
@@ -1,0 +1,15 @@
+import { PaginationDto } from '@/infrastructure/dto/pagination.dto';
+import { ApiPropertyOptional } from '@nestjs/swagger';
+import { IsOptional, IsString } from 'class-validator';
+
+export class SearchBusinessDto extends PaginationDto {
+  @ApiPropertyOptional({ example: 'Pizza Hut' })
+  @IsOptional()
+  @IsString()
+  name?: string;
+
+  @ApiPropertyOptional({ example: 'Pizzeria' })
+  @IsOptional()
+  @IsString()
+  category?: string;
+}
