+++ backend/src/infrastructure/dto/business/update-business.dto.ts
