+++ backend/src/infrastructure/dto/http-error-response.dto.ts
