--- backend/src/modules/category/dto/category-response.dto.ts
@@ -1,20 +1,0 @@
-import { Type } from 'class-transformer';
-import { IsString, ValidateNested } from 'class-validator';
-import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
-
-import { BusinessCategoryResponseDto } from '@/modules/business-category/dto/categories-response.dto';
-
-export class CategoryResponseDto {
-  @ApiProperty({ example: '1' })
-  @IsString()
-  id: string;
-
-  @ApiProperty({ example: 'Peluquería' })
-  @IsString()
-  name: string;
-
-  @ApiPropertyOptional()
-  @ValidateNested({ each: true })
-  @Type(() => BusinessCategoryResponseDto)
-  businessCategories?: BusinessCategoryResponseDto[];
-}
