--- backend/src/modules/user/dto/user-profile-acc.dto.ts
@@ -1,19 +1,0 @@
-import { Exclude, Expose, Type } from 'class-transformer';
-import { ApiProperty } from '@nestjs/swagger';
-
-import { ResponseUserDto } from './user-response.dto';
-import { AccountResponseDto } from '@/modules/account/dto/account-response.dto';
-import { UserProfileResponseDto } from '@/modules/user-profile/dto/userprofile-response.dto';
-
-@Exclude()
-export class UserProfileAccDto extends ResponseUserDto {
-  @ApiProperty({ type: UserProfileResponseDto, nullable: true })
-  @Type(() => UserProfileResponseDto)
-  @Expose()
-  userProfile: UserProfileResponseDto | null;
-
-  @ApiProperty({ type: [AccountResponseDto], nullable: true })
-  @Type(() => AccountResponseDto)
-  @Expose()
-  accounts: AccountResponseDto[] | null;
-}
